# User/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация к серверу User",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/validate-signup": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Проверка данных для регистрации",
                "description": "Метод проверяет корректность данных, предоставленных пользователем при регистрации.",
                "operationId": "77d450a718d147a669a92f45869867b7",
                "requestBody": {
                    "description": "Данные для проверки регистрации",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "user123"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678A!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678A!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Данные прошли проверку",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Данные прошли проверку."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Поле login обязательно для заполнения."
                                        },
                                        "errors": {
                                            "properties": {
                                                "login": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле login обязательно для заполнения."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле email обязательно для заполнения."
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле password обязательно для заполнения."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Internal Server Error",
                                    "error": "SQLSTATE[HY000] [2002] Connection refused"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Выход пользователя из системы",
                "description": "Метод удаляет текущий токен доступа пользователя, завершая сессию.",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Успешный выход"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Internal Server Error",
                                    "error": "SQLSTATE[HY000] [2002] Connection refused"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Получение информации о пользователе",
                "description": "Метод возвращает данные текущего аутентифицированного пользователя и сообщение о действительности токена.",
                "operationId": "f016c0659a1d971af16fc5430fb93b55",
                "responses": {
                    "200": {
                        "description": "Успешное получение данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id_user": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "b9649790-f696-4aa0-9d40-dcfdddfbd9ec"
                                                },
                                                "login": {
                                                    "type": "string",
                                                    "example": "qwerty"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "qwerty@example.com"
                                                },
                                                "created_quizzes_counter": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "taken_quizzes_counter": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-13T21:28:50.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-13T21:28:50.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Токен действителен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Internal Server Error",
                                    "error": "SQLSTATE[HY000] [2002] Connection refused"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите токен в формате: Bearer {your_token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}